// app/admin/_lib/actions/oferta/ofertaDetalle.schemas.ts
export const OfertaDetalleListItemSchema = z.object({
    id: z.string().cuid(),
    ofertaId: z.string().cuid(),
    tituloDetalle: z.string(),
    contenidoExtracto: z.string().optional(),
    orden: z.number().int().nullable().optional(),
    estadoContenido: z.string(),
    updatedAt: z.date(),
    _count: z.object({
        galeriaDetalle: z.number().optional(),
        videoDetalle: z.number().optional(),
        documentosDetalle: z.number().optional(),
    }).optional(),
});
export type OfertaDetalleListItemType = z.infer<typeof OfertaDetalleListItemSchema>;